New setup.py for other machines' compatibility.

+++ setup.py
@@ -3,6 +3,7 @@ import pysam
 import subprocess
 import operator
 import os
+import os.path
 import sys
 
 # Find the ideal -march argument for the system.
@@ -13,7 +14,8 @@ try:
                                   shell=True,
                                   stdout=subprocess.PIPE).stdout.read().strip()
     marchFlag = "-march=%s" % marchValue
-    os.remove("help-dummy.o")
+    if(os.path.isfile("help-dummy.o")):
+        os.remove("help-dummy.o")
 except ImportError:
     print("Error retrieving optimal -march flag. Give up!")
     marchFlag = ""
@@ -34,6 +36,7 @@ compilerList = ["-flto", marchFlag, "-pipe", "-msse2",
 compilerList = ["-O3", "-pipe", marchFlag, "-mfpmath=sse",  # "-flto",
                 "-funroll-loops", "-floop-unroll-and-jam",
                 "-floop-nest-optimize", "-fvariable-expansion-in-unroller"]
+# compilerList = ["-O3", "-pipe", marchFlag]
 
 try:
     from setuptools import setup, Extension
@@ -47,9 +50,7 @@ ext = cythonize('*/*.pyx') + cythonize("*/*.py")
 # If more complex optimizations fail, fall back from line 31 to line 30.
 for x in map(operator.attrgetter("extra_compile_args"), ext):
     # x += ["-Ofast", "-flto", marchFlag, "-pipe",
-    # "-msse2", "-funroll-loops", "-fwhole-program"]
     x += compilerList
-    #  x.extra_compile_args += [marchFlag, "-pipe"]
 
 
 config = {
@@ -72,4 +73,30 @@ config = {
 }
 
 
-setup(**config)
+try:
+    setup(**config)
+except Exception:
+    ext = cythonize('*/*.pyx') + cythonize("*/*.py")
+    compilerList = ["-O3", "-pipe", marchFlag]
+    for x in map(operator.attrgetter("extra_compile_args"), ext):
+        x += compilerList
+    config = {
+        'description': '',
+        'author': 'Daniel Baker',
+        'url': 'https://github.com/ARUP-NGS/BMFTools',
+        'author_email': 'daniel.baker@aruplab.com',
+        'version': '0.0.7.0',
+        'install_requires': ['pysam', 'biopython', 'pudb',
+                             'cython', 'numconv', 'cutadapt', 'lxml'],
+        'packages': ['BMFMain', 'utilBMF', 'MawCluster', 'SECC'],
+        'ext_modules': ext,
+        'include_dirs': [np.get_include()] + pysam.get_include(),
+        'scripts': [],
+        'name': 'BMFTools',
+        'license': 'GNU Affero General Public License, '
+                   'pending institutional approval',
+        'include': 'README.md',
+        'package_data': {'': ['README.md']}
+    }
+    setup(**config)
+
